[CVW-037] Feat, AlertShooter작업동기화 로직 수정
