+++ Projects/Shared/AlertShooter/Sources/Sources/AlertShooter.swift
@@ -7,71 +7,6 @@
 
 import SwiftUI
 
-import I18N
-import CoreUtil
-
-public class DefaultAlertShooter: AlertShooter {
-    private var i18NManager: I18NManager
-    private var languageRepository: LanguageLocalizationRepository
-    
-    public init(i18NManager: I18NManager, languageRepository: LanguageLocalizationRepository) {
-        self.i18NManager = i18NManager
-        self.languageRepository = languageRepository
-        super.init()
-    }
-    
-    public override func shoot(_ model: AlertModel) {
-        Task { [weak self] in
-            guard let self else { return }
-            await container.add(model: model)
-            if alertModel == nil {
-                checkAndShoot()
-            }
-        }
-    }
-    
-    public override func createRO(model: AlertModel) -> AlertRO {
-        let currentLan = i18NManager.getLanguageType()
-        let alertActionROs = model.actions.map { actionModel in
-            let titleText = languageRepository.getString(
-                key: actionModel.titleKey,
-                lanCode: currentLan.lanCode
-            )
-            var titleTextColor: Color!
-            switch actionModel.role {
-            case .normal:
-                titleTextColor = .black
-            case .cancel:
-                titleTextColor = .red
-            }
-            return AlertActionRO(
-                titleText: titleText,
-                titleTextColor: titleTextColor,
-                action: actionModel.action
-            )
-        }
-        
-        // Alert RO
-        var messageText: String = ""
-        if let messageKey = model.messageKey {
-            messageText = languageRepository.getString(
-                key: messageKey,
-                lanCode: currentLan.lanCode
-            )
-        }
-        let alertRO = AlertRO(
-            titleText: languageRepository.getString(
-                key: model.titleKey,
-                lanCode: currentLan.lanCode
-            ),
-            messageText: messageText,
-            actions: alertActionROs
-        )
-        return alertRO
-    }
-}
-
-
 open class AlertShooter: ObservableObject, AlertShootable {
     // State
     @Published var alertModel: AlertRO?
@@ -90,17 +25,21 @@ open class AlertShooter: ObservableObject, AlertShootable {
     
     /// 메세지 스택에서 다음 메세지를 탐색하고 Alert를 표시합니다.
     func checkAndShoot() {
-        Task { [weak self] in
-            guard let self else { return }
-            if let model = await container.getFirst() {
-                // 지연 후 Alert표츌
-                try? await Task.sleep(for: .seconds(0.5))
-                
-                await MainActor.run { [weak self] in
+        // 메인쓰레드 sync실행을 방지하기 위해 사용
+        if let model = container.getFirst() {
+            if Thread.current != Thread.main {
+                DispatchQueue.main.sync { [weak self] in
                     guard let self else { return }
                     alertModel = createRO(model: model)
-                    present = true
                 }
+            } else {
+                alertModel = createRO(model: model)
+            }
+            
+            // 지연 후 Alert표츌
+            DispatchQueue.main.asyncAfter(deadline: .now()+0.5) { [weak self] in
+                guard let self else { return }
+                present = true
             }
         }
     }
